+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.33.3] - 2026-01-23
+
 ### Added
 
 - Continue option now appears on home screen when a session is paused mid-drawing (even if no strokes rendered yet)